Add queue notification protocol

+++ sugar/lib/perq/fsqueue.py
@@ -9,8 +9,10 @@ import os
 import time
 import errno
 import pickle
+import multiprocessing
 
 from sugar.lib.perq.queue import Queue
+from sugar.lib.perq.qexc import QueueEmpty
 from collections import OrderedDict
 import sugar.utils.files
 
@@ -24,16 +26,19 @@ class FSQueue(Queue):
     """
     File-system queue
     """
-    MAX_SIZE = 0xff
-    BUFF = 0xa
+    MAX_SIZE = 0xfff  # Default max size of the queue
+    BUFF = 0xa        # Default buffer
+    POLL = 5          # Poll seconds
 
-    def __init__(self, path, maxsize: int = MAX_SIZE, buff: int = BUFF):
+    def __init__(self, path, maxsize: int = MAX_SIZE, buff: int = BUFF, poll: int = POLL):
         self._queue_path = path
         self._max_size = maxsize
         self._buff_size = buff
         self._buff = OrderedDict()
         self._mutex = False
         self._msgpack = False
+        self._mp_notify = None
+        self._poll = poll
 
         try:
             os.makedirs(self._queue_path)
@@ -53,6 +58,25 @@ class FSQueue(Queue):
         :return:
         """
         self._msgpack = use if msgpack is not None else False
+        return self
+
+    def use_notify(self):
+        """
+        Use queue notification between multi processes.
+
+        Every time when put() is called, internal queue also
+        accepts an object, which indicates that disk has been changed.
+        At that moment get() will re-read the disc store.
+
+        If notify is not used, then disk should be re-read
+        in polling fashion, that might be not always suitable.
+
+        This configuration option also assumes that there is
+        shared queue and it is transferring messages to an end-point.
+
+        :return:
+        """
+        self._mp_notify = multiprocessing.Queue()
         return self
 
     def _lock(self):
@@ -90,10 +114,21 @@ class FSQueue(Queue):
     def __get(self, wait=False):
         self._lock()
         if wait:
-            while self.full():
-                time.sleep(0.01)
+            if self._mp_notify is not None:
+                # Use notification protocol
+                self._mp_notify.get()
+            else:
+                # Poll the disk
+                while True:
+                    if all([True for fname in os.listdir(self._queue_path) if fname.endswith(".xlog")]):
+                        break
+                    time.sleep(self._poll)
 
-        frame_log = os.path.join(self._queue_path, "{}.xlog".format(self._f_dealloc()))
+        xlog = self._f_dealloc()
+        if xlog is None:
+            raise QueueEmpty("Queue is empty")
+
+        frame_log = os.path.join(self._queue_path, "{}.xlog".format(xlog))
         with sugar.utils.files.fopen(frame_log, "rb") as h_frm:
             obj = pickle.load(h_frm)
             os.unlink(frame_log)
@@ -105,7 +140,12 @@ class FSQueue(Queue):
         """
         Deallocate frame
         """
-        return str(list(sorted([int(fname.split(".")[0]) for fname in os.listdir(self._queue_path)]))[0]).zfill(5)
+        try:
+            fn = str(list(sorted([int(fname.split(".")[0]) for fname in os.listdir(self._queue_path)]))[0]).zfill(5)
+        except IndexError:
+            fn = None
+
+        return fn
 
     def _f_alloc(self):
         """
@@ -143,6 +183,10 @@ class FSQueue(Queue):
         frame = self._f_alloc()
         with sugar.utils.files.fopen(os.path.join(self._queue_path, "{}.xlog".format(frame)), "wb") as h_frm:
             pickle.dump(obj, h_frm)
+
+        if self._mp_notify is not None:
+            self._mp_notify.put_nowait(True)
+
         self._unlock()
 
     def qsize(self):
